Split NameResolver and AstToLogical to modules; Add Documentation
